# artifacts/audit/openapi.experimental.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "WorkflowDraftFromUrlAnalysisRequest": {
        "description": "Create a workflow replay draft from URL Intelligence output.",
        "properties": {
          "detected_reason": {
            "default": "",
            "maxLength": 500,
            "title": "Detected Reason",
            "type": "string"
          },
          "original_url": {
            "maxLength": 2048,
            "title": "Original Url",
            "type": "string"
          },
          "selected_start_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Start Url"
          }
        },
        "required": [
          "original_url"
        ],
        "title": "WorkflowDraftFromUrlAnalysisRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "DataForge Engineering",
      "url": "https://github.com/Harshit-sehgal/Scraper/issues"
    },
    "description": "Web extraction backend for supported accessible pages. Supports manual, discovery, and AI-assisted extraction modes with optional browser rendering, content-quality scoring, and responsible rate-limiting. See `docs/API.md` for the full endpoint reference.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/Harshit-sehgal/Scraper/blob/main/LICENSE"
    },
    "title": "DataForge \u2014 Structured Data Extraction API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workflow-drafts/from-url-analysis": {
      "post": {
        "description": "Create a lightweight workflow replay draft from URL Intelligence.\n\nThis is only the Prompt 8 entry point. It does not execute the\nworkflow or attempt to reuse temporary session identifiers.",
        "operationId": "create_workflow_draft_from_url_analysis_api_workflow_drafts_from_url_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDraftFromUrlAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow Draft From Url Analysis",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
    "/api/workflow-drafts/{draft_id}/detect-fields": {
      "post": {
        "description": "Detect form fields for a workflow draft from a local HTML snapshot.",
        "operationId": "detect_workflow_draft_fields_api_workflow_drafts__draft_id__detect_fields_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDraftFieldDetectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect Workflow Draft Fields",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
...
    "/api/workflows": {
      "get": {
        "description": "List workflows with optional filtering.\n\nSupports filtering by status and domain. Results are paginated.",
        "operationId": "list_workflows_api_workflows_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "description": "Create a new workflow from a template or scratch.\n\nRequires at least user role. Owner/org/project are populated from the\nauthenticated context when available.",
        "operationId": "create_workflow_api_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{workflow_id}": {
      "delete": {
        "description": "Delete a workflow permanently.",
        "operationId": "delete_workflow_api_workflows__workflow_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workflow",
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "description": "Get a single workflow by ID.",
        "operationId": "get_workflow_api_workflows__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow",
        "tags": [
          "workflows"
        ]
      },
      "patch": {
        "description": "PATCH alias for partial workflow updates.",
        "operationId": "patch_workflow_api_workflows__workflow_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Workflow",
        "tags": [
          "workflows"
        ]
      },
      "put": {
        "description": "Update an existing workflow. Only provided fields are changed.",
        "operationId": "update_workflow_api_workflows__workflow_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{workflow_id}/preview": {
      "post": {
        "description": "Preview what a workflow would extract without running it.\n\nReturns the first-page extraction result or an error if the workflow\ncannot be previewed (e.g., requires auth, session expired).",
        "operationId": "preview_workflow_api_workflows__workflow_id__preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/WorkflowPreviewRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}